Remade into static, test strings

+++ TawLauncher/UpdateCore.cs
@@ -1,79 +1,97 @@
 ﻿using System;
+using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
 using System.Net;
+using System.Windows;
 
 namespace TawLauncher
 {
-  public class UpdateCore
+  public static class UpdateCore
   {
-    private string _versionFileUrl = null;
-    private string _updateFileUrl = null;
-    private bool _automaticallyUpdate = false;
-    private bool _hideConfigFile = false;
-    private string _exeToRun = null;
-    private bool _runAfterUpdate = false;
-    private bool _keepZip = false;
+    private static string _versionFileUrl = "C:\\Users\\TheAshenWolf\\Desktop\\TawLtest\\version.txt";
+    private static string _updateFileUrl = "C:\\Users\\TheAshenWolf\\Desktop\\TawLtest\\app.zip";
+    private static bool _automaticallyUpdate = true;
+    private static string _exeToRun = "MinecraftClient.exe"; // TODO: Validate if exe
+    private static bool _runAfterUpdate = true;
+    private static bool _keepZip = false;
+    private static bool _keepLauncherOpen = false;
 
-    public Version CheckForUpdates()
+    private static bool _updateAvailable = false;
+
+    public static Version CheckForUpdates()
     {
       Version currentVersion = CheckCurrentVersion();
       Version newVersion = CheckNewVersion();
 
       if (newVersion == null) throw new Exception("Couldn't get updates. Please, check your connection.");
-      if (currentVersion == null) return newVersion;
 
-      if (currentVersion < newVersion) return newVersion;
+      if (currentVersion == null || currentVersion < newVersion)
+      {
+        _updateAvailable = true;
+        if (_automaticallyUpdate) Update();
+        else return newVersion;
+      }
 
-      Directory.Delete(".tmp");
+      if (Directory.Exists("temp")) Directory.Delete("temp", true);
       return null;
     }
 
-    public void Update()
+    public static void Update()
     {
-      if (Directory.Exists("Data")) Directory.Delete("Data");
+      if (!_updateAvailable) return;
+      
+      if (Directory.Exists("Data")) Directory.Delete("Data", true);
       DownloadUpdateFile(_updateFileUrl);
       ZipFile.ExtractToDirectory("Data/Application.zip", "Data/Application");
       if (!_keepZip) File.Delete("Data/Application.zip");
       
-      File.Copy(".tmp/version.txt", "Data/version.txt");
-      Directory.Delete(".tmp");
+      File.Copy("temp/version.txt", "Data/version.txt");
+      Directory.Delete("temp", true);
+      
+      if (_runAfterUpdate) Run();
     }
 
-    private Version CheckCurrentVersion()
+    public static void Run()
+    {
+      Process.Start(Path.GetFullPath("Data/Application/" + _exeToRun));
+      if (!_keepLauncherOpen) Application.Current.Shutdown();
+    }
+
+    private static Version CheckCurrentVersion()
     {
       return !Directory.Exists("Data") ? null : ReadVersionFile();
     }
 
-    private Version CheckNewVersion()
+    private static Version CheckNewVersion()
     {
       if (_versionFileUrl == null) return null;
 
       DownloadVersionFile(_versionFileUrl);
-      return ReadVersionFile(".tmp/version.txt");
+      return ReadVersionFile("temp/version.txt");
     }
 
-    private void CreateTempDirectory()
+    private static void CreateTempDirectory()
     {
-      if (Directory.Exists(".tmp")) Directory.Delete(".tmp");
-      Directory.CreateDirectory(".tmp");
+      if (Directory.Exists("temp")) Directory.Delete("temp", true);
+      Directory.CreateDirectory("temp");
     }
 
-    private void CreateDirectoryTree()
+    private static void CreateDirectoryTree()
     {
       if (Directory.Exists("Data")) return;
       Directory.CreateDirectory("Data");
     }
 
-    private void DownloadVersionFile(string url)
+    private static void DownloadVersionFile(string url)
     {
       CreateTempDirectory();
 
       WebClient webClient = new WebClient();
       try
       {
-        webClient.DownloadFile(url, ".tmp/version.txt");
+        webClient.DownloadFile(url, "temp/version.txt");
       }
       catch
       {
@@ -81,7 +99,7 @@ namespace TawLauncher
       }
     }
 
-    private void DownloadUpdateFile(string url)
+    private static void DownloadUpdateFile(string url)
     {
       CreateDirectoryTree();
 
@@ -96,12 +114,12 @@ namespace TawLauncher
       }
     }
 
-    private Version ReadVersionFile(string path = "Data/version.txt")
+    private static Version ReadVersionFile(string path = "Data/version.txt")
     {
       return new Version(File.ReadLines(path).First());
     }
 
-    private void ReadConfigFile()
+    private static void ReadConfigFile()
     {
       // TODO:
     }
